+++ ta4j-core/src/main/java/org/ta4j/core/BaseTradingRecord.java
@@ -39,6 +39,11 @@ public class BaseTradingRecord implements TradingRecord {
     private static final long serialVersionUID = -4436851731855891220L;
 
     /**
+     * The name of the trading record
+     */
+    private String name;
+
+    /**
      * The recorded orders
      */
     private List<Order> orders = new ArrayList<>();
@@ -93,6 +98,31 @@ public class BaseTradingRecord implements TradingRecord {
 
     /**
      * Constructor.
+     *
+     * @param name the name of the tradingRecord
+     */
+    public BaseTradingRecord(String name) {
+        this(Order.OrderType.BUY);
+        this.name = name;
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param name           the name of the trading record
+     * @param entryOrderType the {@link Order.OrderType order type} of entries in
+     *                       the trading session
+     */
+    public BaseTradingRecord(String name, Order.OrderType orderType) {
+        this(orderType, new ZeroCostModel(), new ZeroCostModel());
+        this.name = name;
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param entryOrderType the {@link Order.OrderType order type} of entries in
+     *                       the trading session
      */
     public BaseTradingRecord(Order.OrderType orderType) {
         this(orderType, new ZeroCostModel(), new ZeroCostModel());
@@ -148,6 +178,11 @@ public class BaseTradingRecord implements TradingRecord {
             Order newOrder = currentTrade.operate(o.getIndex(), o.getPricePerAsset(), o.getAmount());
             recordOrder(newOrder, newOrderWillBeAnEntry);
         }
+    }
+
+    @Override
+    public String getName() {
+        return name;
     }
 
     @Override
@@ -261,7 +296,7 @@ public class BaseTradingRecord implements TradingRecord {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append("BaseTradingRecord:\n");
+        sb.append("BaseTradingRecord: " + name != null ? name : "" + "\n");
         for (Order order : orders) {
             sb.append(order.toString()).append("\n");
         }