+++ ta4j-core/src/main/java/org/ta4j/core/TradingRecord.java
@@ -44,6 +44,11 @@ import static org.ta4j.core.num.NaN.NaN;
 public interface TradingRecord extends Serializable {
 
     /**
+     * @return the name of the TradingRecord
+     */
+    String getName();
+
+    /**
      * @return the current trade
      */
     Trade getCurrentTrade();